refactor(c-56): move getNodeIcon to NodeIcon and drop getTypeLabel
Co-locate node-icon resolution with the NodeIcon component. Remove unused
getTypeLabel + its tests.
