+++ app/utils/__tests__/fileType.test.ts
@@ -1,30 +1,7 @@
-import {
-  Archive,
-  Braces,
-  File,
-  FileSpreadsheet,
-  Folder,
-  Image,
-  Microscope,
-  Table,
-} from "lucide-react";
 import { describe, expect, test } from "vitest";
 
-import { getFileType, getFileTypeIcon, getNodeIcon, getTypeLabel, isImageFile } from "../fileType";
-import { TreeNode } from "~/components/DirectoryView/buildDirectoryTree";
+import { getFileType, getFileTypeIcon, isImageFile } from "../fileType";
 import { formatRegistry } from "~/components/ImageViewer/state/formatRegistry";
-
-function makeNode(overrides: Partial<TreeNode> = {}): TreeNode {
-  return {
-    id: "test",
-    connectionName: "aws-bucket",
-    name: "test",
-    type: "file",
-    pathName: "test",
-    children: [],
-    ...overrides,
-  };
-}
 
 describe("getFileType", () => {
   test("identifies OME-TIFF files (takes precedence over TIFF)", () => {
@@ -150,10 +127,6 @@ describe("query-string and fragment handling", () => {
     expect(isImageFile("photo.jpg?expires=tomorrow")).toBe(true);
     expect(isImageFile("data.csv?range=0-100")).toBe(false);
   });
-
-  test("getTypeLabel fallback strips suffix before uppercase", () => {
-    expect(getTypeLabel(makeNode({ name: "weird.xyz?sig=abc" }))).toBe("XYZ");
-  });
 });
 
 describe("getFileTypeIcon", () => {
@@ -174,93 +147,5 @@ describe("getFileTypeIcon", () => {
   test("returns File for unknown types", () => {
     expect(getFileTypeIcon("unknown.xyz")).toBe("File");
     expect(getFileTypeIcon("")).toBe("File");
-  });
-});
-
-describe("getNodeIcon", () => {
-  test("returns Archive for bucket nodes", () => {
-    expect(getNodeIcon(makeNode({ type: "bucket" }))).toBe(Archive);
-  });
-
-  test("returns Folder for directory nodes", () => {
-    expect(getNodeIcon(makeNode({ type: "directory" }))).toBe(Folder);
-  });
-
-  test("returns Microscope for .ome.tif files", () => {
-    expect(getNodeIcon(makeNode({ name: "sample.ome.tif" }))).toBe(Microscope);
-  });
-
-  test("returns Microscope for .ome.tiff files", () => {
-    expect(getNodeIcon(makeNode({ name: "sample.ome.tiff" }))).toBe(Microscope);
-  });
-
-  test("returns Image for .png files", () => {
-    expect(getNodeIcon(makeNode({ name: "photo.png" }))).toBe(Image);
-  });
-
-  test("returns Image for .jpg files", () => {
-    expect(getNodeIcon(makeNode({ name: "photo.jpg" }))).toBe(Image);
-  });
-
-  test("returns FileSpreadsheet for .csv files", () => {
-    expect(getNodeIcon(makeNode({ name: "data.csv" }))).toBe(FileSpreadsheet);
-  });
-
-  test("returns Table for .parquet files", () => {
-    expect(getNodeIcon(makeNode({ name: "data.parquet" }))).toBe(Table);
-  });
-
-  test("returns Braces for .json files", () => {
-    expect(getNodeIcon(makeNode({ name: "config.json" }))).toBe(Braces);
-  });
-
-  test("returns File for unknown extensions", () => {
-    expect(getNodeIcon(makeNode({ name: "readme.md" }))).toBe(File);
-  });
-
-  test("returns File for files without extension", () => {
-    expect(getNodeIcon(makeNode({ name: "Makefile" }))).toBe(File);
-  });
-});
-
-describe("getTypeLabel", () => {
-  test("returns 'Bucket' for bucket nodes", () => {
-    expect(getTypeLabel(makeNode({ type: "bucket" }))).toBe("Bucket");
-  });
-
-  test("returns 'Folder' for directory nodes", () => {
-    expect(getTypeLabel(makeNode({ type: "directory" }))).toBe("Folder");
-  });
-
-  test("returns 'OME-TIFF' for .ome.tif files", () => {
-    expect(getTypeLabel(makeNode({ name: "sample.ome.tif" }))).toBe("OME-TIFF");
-  });
-
-  test("returns 'CSV' for .csv files", () => {
-    expect(getTypeLabel(makeNode({ name: "data.csv" }))).toBe("CSV");
-  });
-
-  test("returns 'Parquet' for .parquet files", () => {
-    expect(getTypeLabel(makeNode({ name: "data.parquet" }))).toBe("Parquet");
-  });
-
-  test("returns 'PNG' for .png files", () => {
-    expect(getTypeLabel(makeNode({ name: "image.png" }))).toBe("PNG");
-  });
-
-  test("returns 'JPEG' for .jpg files", () => {
-    expect(getTypeLabel(makeNode({ name: "photo.jpg" }))).toBe("JPEG");
-  });
-
-  test("returns 'NDJSON' for .ndjson files", () => {
-    expect(getTypeLabel(makeNode({ name: "stream.ndjson" }))).toBe("NDJSON");
-  });
-
-  test("returns uppercase extension for unknown types", () => {
-    expect(getTypeLabel(makeNode({ name: "readme.md" }))).toBe("MD");
-  });
-
-  test("returns 'File' for files without extension", () => {
-    expect(getTypeLabel(makeNode({ name: "Makefile" }))).toBe("File");
   });
 });
